% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_desc.R
\name{reorder_desc}
\alias{reorder_desc}
\title{Reorder levels of a factors in descending order}
\usage{
reorder_desc(x, y)
}
\arguments{
\item{x}{a factor (or character vector)}

\item{y}{a vector of the same length as x, based on which levels of x will be ordered.
If y is a character vector, it will be sorted in reverse alphabetic order and x ordered accordingly.}
}
\description{
Descending version of the `reorder` function. Reorders levels of a factor based on descending order of values of a second variable.
}
\examples{
x <- c("a", "d", "h", "g", "m")
y <- c(1, 3, 9, 7, 6)
reorder_desc(x, y)

}
